correct of memcmp

+++ libft/ft_memcmp.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-size_t	ft_memcmp(const void *s1, const void *s2, size_t n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	size_t			i;
 	unsigned char	*s1_temp;
